Move ConsentDAO to common
