+++ dpc-aggregation/src/main/java/gov/cms/dpc/aggregation/service/ConsentServiceImpl.java
@@ -1,6 +1,6 @@
 package gov.cms.dpc.aggregation.service;
 
-import gov.cms.dpc.aggregation.jdbi.ConsentDAO;
+import gov.cms.dpc.common.jdbi.ConsentDAO;
 import gov.cms.dpc.common.consent.entities.ConsentEntity;
 import gov.cms.dpc.fhir.DPCIdentifierSystem;
 import gov.cms.dpc.fhir.converters.entities.ConsentEntityConverter;
