+++ dpc-common/src/test/java/gov/cms/dpc/common/jdbi/ConsentDAOUnitTest.java
@@ -1,8 +1,9 @@
-package gov.cms.dpc.consent.jdbi;
+package gov.cms.dpc.common.jdbi;
 
 import gov.cms.dpc.common.consent.entities.ConsentEntity;
 import gov.cms.dpc.common.consent.entities.OptOutFileEntity;
 import gov.cms.dpc.common.hibernate.consent.DPCConsentManagedSessionFactory;
+import gov.cms.dpc.common.jdbi.ConsentDAO;
 import gov.cms.dpc.testing.AbstractMultipleDAOTest;
 import gov.cms.dpc.testing.utils.MBIUtil;
 import org.junit.jupiter.api.BeforeEach;
@@ -21,7 +22,7 @@ class ConsentDAOUnitTest extends AbstractMultipleDAOTest {
 	private ConsentDAO consentDAO;
 
 	@BeforeEach
-	public void setup() {
+    void setup() {
 		consentDAO = new ConsentDAO(new DPCConsentManagedSessionFactory(db.getSessionFactory()));
 	}
 
@@ -29,9 +30,7 @@ class ConsentDAOUnitTest extends AbstractMultipleDAOTest {
 	void test_ConsentDAO_writes_consent() {
 		ConsentEntity consentEntity = createConsentEntity();
 
-		db.inTransaction(() -> {
-			consentDAO.persistConsent(consentEntity);
-		});
+		db.inTransaction(() -> consentDAO.persistConsent(consentEntity));
 
 		ConsentEntity returnedConsent = consentDAO.getConsent(consentEntity.getId()).get();
 		assertEquals(consentEntity.getMbi(), returnedConsent.getMbi());
@@ -64,6 +63,27 @@ class ConsentDAOUnitTest extends AbstractMultipleDAOTest {
 		assertEquals(1, returnedList.size());
 		assertEquals(consentEntity.getId(), returnedList.get(0).getId());
 	}
+
+    @Test
+    void test_ConsentDAO_findByMbis() {
+        ConsentEntity entity1 = createConsentEntity();
+        entity1.setMbi("fake_mbi_1");
+        ConsentEntity entity2 = createConsentEntity();
+        entity2.setMbi("fake_mbi_2");
+        ConsentEntity entity3 = createConsentEntity();
+        entity3.setMbi("fake_mbi_3");
+
+        db.inTransaction(() -> {
+            consentDAO.persistConsent(entity1);
+            consentDAO.persistConsent(entity2);
+            consentDAO.persistConsent(entity3);
+        });
+
+        List<String> mbis = List.of("fake_mbi_1", "fake_mbi_2");
+        List<ConsentEntity> returnedList = consentDAO.findByMbis(mbis);
+        assertEquals(2, returnedList.size());
+        assertEquals(mbis, returnedList.stream().map(ConsentEntity::getMbi).toList());
+    }
 
 	private ConsentEntity createConsentEntity() {
 		ConsentEntity consentEntity = new ConsentEntity();
